refactor : App 구조개선 (Home,router 분리)
