+++ src/App.tsx
@@ -1,34 +1,8 @@
-import { createBrowserRouter, RouterProvider } from 'react-router';
-import Navigation from './components/Navigation/Navigation';
-import FriendSelector from './components/FriendSelector/FriendSelector';
-import CategorySection from './components/CategorySection/CategorySection';
-import MessageSection from './components/MessageSection/MessageSection';
-import RankingSection from './components/RankingSection/RankingSection';
-import Login from './pages/Login/Login';
-
-const HomePage = () => (
-  <>
-    <Navigation />
-    <FriendSelector />
-    <CategorySection />
-    <MessageSection />
-    <RankingSection />
-  </>
-);
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <HomePage />
-  },
-  {
-    path: '/login',
-    element: <Login />
-  }
-]);
+import { RouterProvider } from 'react-router-dom';
+import routes from './routes/router';
 
 function App() {
-  return <RouterProvider router={router} />;
+  return <RouterProvider router={routes} />;
 }
 
-export default App
+export default App;
