+++ src/routes/router.tsx
@@ -1,0 +1,16 @@
+import { createBrowserRouter } from 'react-router-dom';
+import Home from '../pages/Home/Home';
+import Login from '../pages/Login/Login';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Home />,
+  },
+  {
+    path: '/login',
+    element: <Login />,
+  },
+]);
+
+export default router;
